msm: ais: memmgr: Avoid TOCTOU buffer access during multiple use of same fd
Fd is a user-accessible value, referring it multiple times
leads to TOCTOU issues. Dma_buf can be freed after the 1st
use of fd and userspace can create another dma_buf but with
same fd. In such scenario, during 2nd use of fd, we may get
a different dma_buf with different length. To avoid this, we
can use same dma_buf instead of retrieving it twice using same
fd. In this change FD is accessed only once in syscall.

Change-Id: I0a85e348e200b49ffd37608915d2182e8add5ecc
Signed-off-by: anshuk <quic_anshuk@quicinc.com>