msm: camera: sensor: avoid cascading effect in no crm apply
If sensor packet is delayed from umd to kmd, and for isp,  that packet is
already available isp will apply that packet and send apply to sensor
however sensor would not be able to apply the packet as sensor doesnot
have that packet available in kmd. So from that point onwards sensor
will be out of sync with isp and would never be corrected.

The above scenario can happen multiple times in a single session
and the delta between isp and sensor request id applied keeps increasing
in each instance of this event.

To correct this out of sync we need to handle sensor apply based on isp
applied request id. This change implements the logic to
correct the above mentioned out of sync scenario based on isp applied
request id.

CRs-Fixed: 3337452
Change-Id: I402b09c28c3f270cfdb9aa694b853742504da5f0
Signed-off-by: Tony Lijo Jose <quic_tjose@quicinc.com>