msm: camera: isp: Updates to CSID IRQ handling
When streaming off, issue a reset and then maskout all CSID
IRQs together as opposed to manipulating the mask bit by bit.
There is no atomic way to maskout all IRQs, in the process of
masking out IRQ bits an IRQ could still fire. This change ensures
we maskout the IRQs in a clean way post that unsubscribe for
those events from the IRQ controller. The change also ensures
the rx input is cut off prior to issuing a SW reset, this is
to rule out any spurious activity in HW caused due to incoming
data. At the time of the SW reset we are certain, we are not
in the middle of processing frames.

CRs-Fixed: 3336496
Change-Id: I1f714d290d31bb744809b2df9e2a4245b59641fe
Signed-off-by: Karthik Anantha Ram <quic_kartanan@quicinc.com>
(cherry picked from commit a3adb7a4420fe55d5ed3b1e1e6cb6fd90680d3e4)