msm: camera: isp: Add check the back to back sof happen
When the bubble happen, and then there is back to back sof
handle happen. For the rdi and stats only context and no crm,
after bubble happen, state will be bubbled, and then handle
the sof, currently will signal the active request with
fence error to umd, and umd will unmap the request buffer,
but actually the request buffer done is not coming, which maybe
cause smmu page fault issue. This commit add check the back
to back sof handle, if so do not signal any fences.

CRs-Fixed: 3513868
Change-Id: I14c68f4c28fe95fe2e5401f191ea09758e7e10a8
Signed-off-by: zhuo <quic_zhuo@quicinc.com>