msm: camera: isp: Config hardware in work queue
Currently requests for offline IFE are applied from tasklet
context or from an ioctl thread within a critical section
protected by a spinlock. In both of these cases, preemtion is
disabled but CDM requires mutex locks during this operation.
Therefore a work queue is added in this change to handle
apply request for offline IFE.

Change-Id: I22fcbc61801a0c217a27bf3aa9b36641f9b16c67
Signed-off-by: Venkat Chinta <vchinta@codeaurora.org>
Signed-off-by: Zun Qiao <quic_zqiao@quicinc.com>