msm: ais: fix cci read after NACK
Fix an issue where cam_cci_read_bytes() fails after
a previous I2C transaction failed with NACK error.
This is the case when the I2C client from the previous
transaction does not ACK the request bytes from master.
Reason for this could be unresponsive I2C client or
its disconnection from the I2C Bus.

Root cause of the issue is that rd_done completion which
signals finish of a read transaction via IRQ status is
not reset  before a read transaction.
This lead to a race condition between the actual "read
done" IRQ in cam_cci_irq() and the process context thread
waiting in cam_cci_read().
The wait_for_completion_timeout() within this function
returns immediately since completion is still marked as
done from previous transaction.
In some cases the operation is actually done and no
issue is seen. But in other cases the operation is
still in progress which leads to cam_cci_read()
being unable to process the received bytes sent by
I2C client.

The fix for the root cause is to reset the rd_done
completion such that cam_cci_read() blocks until
operation is really completed.

Change-Id: I444974085ffb0cce2d2ec07428ab44530cff390e
Signed-off-by:  Sudhakara Rao Kimidi <skimidi@qti.qualcomm.com>
Signed-off-by: Vandana Jain <quic_vandjain@quicinc.com>