msm: camera: sensor: Resolving boot crash
The kernel will do might_sleep check when we down a semaphore,
if the might_sleep found it is in atomic context, it will
eventually call BUG() to raise kernel panic because sleep is
not permitted in atomic context. What we doing now is getting
a spink lock before down a semaphore, this is possible to make
might_sleep raise a kernel panic because semaphore might sleep
in spin lock. So now we have removed the mutex and spinlock
in the function.

CRs-Fixed: 3580578
Change-Id: I244a3123e181d64197593b21758e40b00db680d5
Signed-off-by: Rajat Goel <quic_rajgoe@quicinc.com>