msm: camera: smmu: Use dmabuf inode num as unique identifier
'fd' can be closed in user mode driver, though corresponding
dmabuf is still not freed if someone still holding a refCount
on that dmabuf, say camera smmu driver holding a refCOunt and
has that in its mapping info table. Then the same 'fd' can be
allotted for a different buffer in a new allocation. In such
cases, if we use 'fd' as identifier in camera smmu driver to
check if the buffer is already mapped based on mapping info
table, we mistakenly think the new buffer is already mapped
which is not the case.

DMA buf's inode num is an unique identifier that represents
a buffer. Use this along with 'fd'in camera smmu driver to
see if the given buffer is already mapped.

CRs-Fixed: 3805063
Change-Id: Ie660be804f2c3c3483d2824feeab6fceedb017cc
Signed-off-by: Shivakumar Malke <quic_smalke@quicinc.com>