msm: camera: isp: Handle HW release in perport usecase
This change fixes prepare hw update failure during
config dev in perport usecase due to HW release.
When one of the perport cameras is closed, flush
happens and HW is released. This change make sure
that HW is released only when all the perport cameras
in a group are closed.

CRs-Fixed: 3854122
Change-Id: I51c9f1ce75a50cedc041bc9a2365e6b18f1dd131
Signed-off-by: Karthik Dillibabu <quic_kard@quicinc.com>